# ⚡️ Improve `SequenceSet#xor` performance by ~2x