@nevans 2025-04-29
I didn't have proper tests over `#xor` and had accidentally pasted in an alternate implementation for `#union`!  D'oh!